requirements.txt
Markdown==3.5.1
PyYAML==6.0.1
titlecase==2.4.1